requirements.txt
pandas~=2.2.3
mysql-connector-python~=9.2.0
numpy~=2.2.3
ttkthemes~=3.2.2
openpyxl~=3.1.5
aiohttp~=3.11.13
networkx~=3.4.2
scipy~=1.15.2
plotly~=6.0.0
scikit-learn~=1.6.1
pillow~=11.1.0